Add files via upload
This node allows the text box to execute python code. Print your string value to assign it to the output. If the output is a number, it will also be available on the INT or FLOAT outputs.
There is one special upper case constant you can reference from the code box. PYTHON_CODE_BOX_SEED. This value holds the node's seed value. You can convert this value to an input and drive it by other methods.